Move GenerateManfiest.run() to generate()
