+++ src/commands/addons/admin/manifest/generate.ts
@@ -7,7 +7,7 @@ import {prompt} from 'inquirer'
 import {generate as generateString} from 'randomstring'
 
 import AdminBase from '../../../../admin-base'
-import {GenerateManifest} from '../../../../manifest'
+import {ManifestInterface} from '../../../../manifest'
 
 export default class Generate extends AdminBase {
   static description = 'generate a manifest template'
@@ -37,7 +37,7 @@ The file has been saved!`,
 
     // prompts for manifest
     const promptAnswers = await this.askQuestions(flags, regions)
-    const manifest = GenerateManifest.run(promptAnswers)
+    const manifest = this.generate(promptAnswers)
     await this.writeManifest(manifest)
   }
 
@@ -138,5 +138,42 @@ The file has been saved!`,
       }
       this.log(`The file ${color.green('addon_manifest.json')} has been saved!`)
     })
+  }
+
+  private generate(data: any = {}): ManifestInterface {
+    let manifest: ManifestInterface = {
+      id: 'myaddon',
+      api: {
+        config_vars_prefix: 'MYADDON',
+        config_vars: [
+          'MYADDON_URL'
+        ],
+        password: 'CHANGEME',
+        sso_salt: 'CHANGEME',
+        regions: ['us', 'eu'],
+        requires: [],
+        production: {
+          base_url: 'https://myaddon.com/heroku/resources',
+          sso_url: 'https://myaddon.com/sso/login'
+        },
+        test: {
+          // tslint:disable-next-line:no-http-string
+          base_url: 'http://localhost:4567/heroku/resources',
+          // tslint:disable-next-line:no-http-string
+          sso_url: 'http://localhost:4567/sso/login'
+        },
+        version: '3'
+      },
+      name: 'MyAddon',
+    }
+
+    manifest.id = data.id || manifest.id
+    manifest.api.config_vars_prefix = (data.id ? data.id.toUpperCase() : manifest.api.config_vars_prefix)
+    manifest.api.config_vars = (data.id ? [`${data.id.toUpperCase()}_URL`] : manifest.api.config_vars)
+    manifest.api.password = data.password || manifest.api.password
+    manifest.api.sso_salt = data.sso_salt || manifest.api.sso_salt
+    manifest.api.regions = data.regions || manifest.api.regions
+    manifest.name = data.name || manifest.name
+    return manifest
   }
 }
